requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.6
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-js-asset==2.1.0
django-taggit==5.0.0
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
